+++ Sharkingbird.Jarvis.Core.Models/Transaction.cs
@@ -7,7 +7,7 @@ namespace Sharkingbird.Jarvis.Core.Models
 {
   public sealed class Transaction
   {
-    public Transaction(Guid id, double amount, string description, DateTimeOffset appliedOn, RecurringTransactionNameValueType? recurringTransactionName)
+    public Transaction(Guid id, decimal amount, string description, DateTimeOffset appliedOn, RecurringTransactionNameValueType? recurringTransactionName)
     {
       Id = id;
       Amount = amount;
@@ -17,7 +17,7 @@ namespace Sharkingbird.Jarvis.Core.Models
     }
     public Guid Id { get; }
     public RecurringTransactionNameValueType? RecurringTransactionName { get; }
-    public double Amount { get; }
+    public decimal Amount { get; }
     public string Description { get; }
     public DateTimeOffset AppliedOn { get; }
   }
