+++ Sharkingbird.Jarvis.UnitTests/BudgetTests.cs
@@ -10,7 +10,7 @@ namespace Sharkingbird.Jarvis.UnitTests
   {
     [TestCase("RecurringTransaction",9.99,"Budget",RecurringRateEnum.Weekly)]
     public void Applying_Recurring_Transaction_To_Empty_Budget_Applies_Transaction(string recurringTransactionName,
-      double amount, string budgetName, RecurringRateEnum rate)
+      decimal amount, string budgetName, RecurringRateEnum rate)
     {
       var budget = new Budget(budgetName, new Transaction[0]);
 
@@ -25,7 +25,7 @@ namespace Sharkingbird.Jarvis.UnitTests
 
     [TestCase("RecurringTransaction", 9.99, "Budget", RecurringRateEnum.Weekly)]
     public void Applying_Recurring_Transaction_Twice_Only_Applies_Transaction_Once(string recurringTransactionName,
-      double amount, string budgetName, RecurringRateEnum rate)
+      decimal amount, string budgetName, RecurringRateEnum rate)
     {
 
       var budget = new Budget(budgetName, new Transaction[0]);
@@ -42,7 +42,7 @@ namespace Sharkingbird.Jarvis.UnitTests
 
     [TestCase("RecurringTransaction", 9.99, "Budget", RecurringRateEnum.Weekly)]
     public void Applying_Recurring_Transaction_When_Transaction_Exists_Does_Not_Apply_New_Transaction(string recurringTransactionName,
-      double amount, string budgetName, RecurringRateEnum rate)
+      decimal amount, string budgetName, RecurringRateEnum rate)
     {
 
       var budget = new Budget(budgetName, new[] {
@@ -58,7 +58,7 @@ namespace Sharkingbird.Jarvis.UnitTests
 
     [TestCase("RecurringTransaction", 9.99, "Budget")]
     public void Applying_Weekly_Transaction_8_Days_Ago_Applies_New_Transaction_Once(string recurringTransactionName,
-      double amount, string budgetName)
+      decimal amount, string budgetName)
     {
       var eightDaysAgo = DateTimeOffset.Now - TimeSpan.FromDays(8);
 
@@ -70,13 +70,13 @@ namespace Sharkingbird.Jarvis.UnitTests
       budget.ApplyRecurringTransactions(new[] { recurringTransaction });
 
 
-      Assert.AreEqual(2.0 * amount, budget.Balance);
+      Assert.AreEqual(2 * amount, budget.Balance);
       Assert.AreEqual(1, budget.NewTransactions.Count());
     }
 
     [TestCase("RecurringTransaction", 9.99, "Budget")]
     public void Applying_Weekly_Transaction_15_Days_Ago_Applies_New_Transaction_Twice(string recurringTransactionName,
-      double amount, string budgetName)
+      decimal amount, string budgetName)
     {
       var fifteenDaysAgo = DateTimeOffset.Now - TimeSpan.FromDays(15);
 
@@ -88,7 +88,7 @@ namespace Sharkingbird.Jarvis.UnitTests
       budget.ApplyRecurringTransactions(new[] { recurringTransaction });
 
 
-      Assert.AreEqual(3.0 * amount, budget.Balance);
+      Assert.AreEqual(3 * amount, budget.Balance);
       Assert.AreEqual(2, budget.NewTransactions.Count());
     }
   }
